+++ index.html
@@ -33,8 +33,8 @@
     <section class="add-book">
     <h1>Add a new book</h1>
         <form>
-            <input class="book-title" type="text" name="tite" id="" placeholder="Title">
-            <input class="author-name" type="text" name="author" id="" placeholder="Author">
+            <input class="book-title" type="text" name="tite" id="" placeholder="Title" required>
+            <input class="author-name" type="text" name="author" id="" placeholder="Author" required>
             <button class="add-btn" type="submit">Add</button>
         </form>
     </section>
