+++ modules/book-class.js
@@ -1,5 +1,4 @@
-// eslint-disable-next-line import/prefer-default-export
-export class Book {
+export default class Book {
   constructor(title, author) {
     this.title = title;
     this.author = author;
