+++ modules/clear-fields.js
@@ -1,5 +1,5 @@
-// eslint-disable-next-line import/prefer-default-export
-export const clearFields = () => {
+const clearFields = () => {
   document.querySelector('.book-title').value = '';
   document.querySelector('.author-name').value = '';
 };
+export default clearFields;
